adjust testcases
